Refactor NativeMethods
We should not expose C interfaces beyond where they are needed. To reduce
usage errors, I've wrapped all functions that acceptor or return an IntPtr
with new types. This eliminates programmer error where they send a pointer
of the wrong type to NativeMethod.XXX. Also new is the enumeration would
which pre-enumerates all devices so the caller doesn't have to marshal
the list of pointers into their own list and then remember to call free.
The implementation of wrapper class now call HidInit and the dispose method
call HidExit.