Logo API added
This adds logo storage and printing support. The logos will
be resized and dithered as configured by LogoStorageConfig
then stored in external flash. Logos can be recalled using
the PrintLogo(index) function. The logo writing process is
similar to flash updates in that it uses the BinaryFile
class for data input and commnucates progress via the
IProgressMonitor interface.