Update project layout to support runtimes/rid pattern
There are conflicting nuget docs but basically this folder layout
with the runtimes/{rid}/native structure *should* make nuget happy.
Initial testing shows that the dlls are correctly picked up and packed
into the nuget. The remaining task is how to get the correct version
copied back to the consumers output directory.