Bug fix in targets file
Was accidentally copying the x64 version instead of the x86 version
to the output directory.
Create two new solutions folders for some of the random tools and files
in this project. Fix warning about operator precedence in targets file

Resolve issue #6

This resolves the docfx documentation metadata issue that broke online docs.

Only deploy docs if not on a PR check

Add net461 target for older versions of visual studio
VS 2013 does not acceptor netstandard2.0 as a valid target even with net 4.6.1 installed. Adding this tag to the nuget fixes the issue for these users.

Add IsDeviceReady property to ReliancePrinter
This property gives the client the ability to reliably check on the status of the low-level comms interface. Also, for the CLI force-copy the hidapi.dll to the output directory to make debugging easier.