Multiprinter support and extensible ReliancePrinter class
This adds support for connecting to a ReliancePrinter with an additional constraint on the device path. If the path is not specified, the first match is returned which was the original behavior. If path is specified, the handle will only be returned if the DevicePath matches what was specified in the Open method. Add function count attached reliance printers.

With multiple devices, we need to control who is opening and closing which device path. The NativeMethods class now tracks the owned paths globally. An owned path is a path that has been successfully opened by any instance of NativeMethods. Two helpers were added to support this new feature. IsPathOwned checks is the path is currently open in any instance. If this returns true, no attempt should be made to open the path. The second helper is IsPathOwnedByThis checks if the path is owned by the current NativeMethods instance. A path should only be closed by its owner.

Use Auto character set for error string retrieval in NativeMethods. Add more trace level logging to NativeMethods class.

Expose WriteRaw as a protected function

Add new filter features to CLI sample app. Add new multitest mode to CLI app.

Add verbose explanation to DLLNotFoundException. For the unitiated, this may be a weird error. Explain what exactly went wrong and how to fix it with appropriate links to Github.

Cleanup unit tests