Widen Newtonsoft.Json supported version window
Support and version of Newtonsoft.JSON down to 9.0.1

Support down to 9.0.1 because that is the minimum for dotnetcore
support. Set the upper limit 12.0.1.

Also upgraded liblog so we can have less building noise. The internal
classes were not documented on the previous version which generated
a ton of warnings at build time.